requirements.txt
langchain
langchain_openai
streamlit
python-dotenv
